+++ src-docs/src/views/bottom_bar/bottom_bar_displacement.js
@@ -1,0 +1,58 @@
+import React, { useState } from 'react';
+
+import {
+  EuiBottomBar,
+  EuiButtonGroup,
+  EuiButtonEmpty,
+  EuiFlexGroup,
+  EuiFlexItem,
+} from '../../../../src/components';
+
+export default () => {
+  const [toggleIdSelected, setToggleIdSelected] = useState(null);
+
+  const toggleButtons = [
+    {
+      id: 'bottomBarStandard',
+      label: 'Show bottom bar',
+    },
+    {
+      id: 'bottomBarWithoutAffordForDisplacement',
+      label: 'Show bottom bar (without affordForDisplacement behavior)',
+    },
+  ];
+
+  const onChange = (optionId) => {
+    setToggleIdSelected(optionId);
+  };
+
+  return (
+    <div>
+      <EuiButtonGroup
+        legend="Bottom Bar demo toggle buttons group"
+        type="single"
+        buttonSize="m"
+        options={toggleButtons}
+        idSelected={toggleIdSelected}
+        onChange={(id) => onChange(id)}
+      />
+
+      {toggleIdSelected && (
+        <EuiBottomBar
+          affordForDisplacement={toggleIdSelected === 'bottomBarStandard'}>
+          <EuiFlexGroup justifyContent="flexEnd">
+            <EuiFlexItem grow={false}>
+              <EuiButtonEmpty
+                onClick={() => setToggleIdSelected(null)}
+                color="ghost"
+                size="s"
+                iconType="cross">
+                close
+              </EuiButtonEmpty>
+            </EuiFlexItem>
+          </EuiFlexGroup>
+        </EuiBottomBar>
+      )}
+    </div>
+  );
+};